feat: Improves 2_pc rendering. Better colormaps.
- Enhances the colormaps using a value gradient in a HSV model. The
gradient allows to percieve the 3D space better than the previous
colormaps, and the usage of the HSV model not only allows for computing
this gradient but also makes it easy to choose a pair of distinguishable
colors.
- Uses different materials for the two pointclouds. It should improve
the independence of each pointcloud visually, but it's hard to see, the
colormaps helped more.